requirements.txt
streamlit==1.50.0
numpy==2.3.3
matplotlib==3.10.7
sympy==1.14.0
pillow==11.3.0
pytesseract==0.3.13
opencv-python==4.10.0.84
pandas==2.3.3
plotly==5.24.1
scipy==1.15.2
